+++ client/src/lib/actions/taxi-calls.ts
@@ -1,6 +1,6 @@
 import Logger from "../Logger";
 import { Dispatch } from "react";
-import { CREATE_TAXI_CALL, SET_MESSAGE, GET_TAXI_CALLS } from "../types";
+import { CREATE_TAXI_CALL, SET_MESSAGE, GET_TAXI_CALLS, END_TAXI_CALL } from "../types";
 import { handleRequest, isSuccess } from "../functions";
 import socket from "../socket";
 import Message from "../../interfaces/Message";
@@ -43,5 +43,20 @@ export const createTaxiCall = (data: object) => async (dispatch: Dispatch<IDispa
     }
   } catch (e) {
     Logger.error(CREATE_TAXI_CALL, e);
+  }
+};
+
+export const endTaxiCall = (id: string) => async (dispatch: Dispatch<IDispatch>) => {
+  try {
+    const res = await handleRequest(`/taxi-calls/${id}`, "DELETE");
+
+    if (isSuccess(res)) {
+      dispatch({
+        type: END_TAXI_CALL,
+        calls: res.data.calls,
+      });
+    }
+  } catch (e) {
+    Logger.error(END_TAXI_CALL, e);
   }
 };