+++ server/src/routes/management.ts
@@ -10,6 +10,11 @@ import Logger from "../lib/Logger";
 import Officer from "../interfaces/Officer";
 const router: Router = Router();
 
+interface Item {
+  value: string;
+  label: string;
+}
+
 /* Cad settings */
 router.put("/cad-settings", useAuth, async (req: IRequest, res: Response) => {
   const user = await processQuery<IUser[]>("SELECT `rank` from `users` WHERE `id` = ?", [
@@ -147,6 +152,71 @@ router.get("/citizens", useAuth, useAdminAuth, async (_req: IRequest, res: Respo
 
   return res.json({ citizens, status: "success" });
 });
+
+router.get("/expungement-requests", useAuth, useAdminAuth, async (_req, res: Response) => {
+  const requests = await processQuery("SELECT * FROM `court_requests`");
+
+  return res.json({
+    status: "success",
+    requests: requests.map((request: any) => {
+      request.warrants = JSON.parse(request.warrants);
+      request.arrestReports = JSON.parse(request.arrest_reports);
+      request.tickets = JSON.parse(request.tickets);
+      return request;
+    }),
+  });
+});
+
+router.put(
+  "/expungement-requests/:requestId/:type",
+  useAuth,
+  useAdminAuth,
+  async (req: IRequest, res: Response) => {
+    const { requestId, type } = req.params;
+
+    switch (type) {
+      case "accept": {
+        // value to remove from the citizen
+        const { warrants, arrestReports, tickets } = req.body;
+
+        warrants.forEach(async (warrant: Item) => {
+          await processQuery("DELETE FROM `warrants` WHERE `id` = ?", [warrant.value]);
+        });
+        arrestReports.forEach(async (arr: Item) => {
+          await processQuery("DELETE FROM `arrest_reports` WHERE `id` = ?", [arr.value]);
+        });
+        tickets.forEach(async (ticket: Item) => {
+          await processQuery("DELETE FROM `leo_tickets` WHERE `id` = ?", [ticket.value]);
+        });
+
+        await processQuery("DELETE FROM `court_requests` WHERE `id` = ?", [requestId]);
+        break;
+      }
+      case "decline": {
+        await processQuery("DELETE FROM `court_requests` WHERE `id` = ?", [requestId]);
+        break;
+      }
+      default: {
+        return res.json({
+          error: "invalid type",
+          status: "error",
+        });
+      }
+    }
+
+    const updated = await processQuery("SELECT * FROM `court_requests`");
+
+    return res.json({
+      status: "success",
+      requests: updated.map((request: any) => {
+        request.warrants = JSON.parse(request.warrants);
+        request.arrestReports = JSON.parse(request.arrest_reports);
+        request.tickets = JSON.parse(request.tickets);
+        return request;
+      }),
+    });
+  }
+);
 
 router.delete("/citizens/:id", useAuth, useAdminAuth, async (req: IRequest, res: Response) => {
   const { id } = req.params;